requirements.txt
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-secret-manager
google-cloud-bigquery
google-api-python-client
google-adk
